requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.7.6.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
